+++ core/primitives/CMakeLists.txt
@@ -8,6 +8,7 @@ add_subdirectory(block)
 add_subdirectory(chain)
 add_subdirectory(cid)
 add_subdirectory(piece)
+add_subdirectory(resources)
 add_subdirectory(rle_bitset)
 add_subdirectory(sector)
 add_subdirectory(sector_file)
