+++ core/primitives/resources/resources.hpp
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef CPP_FILECOIN_CORE_PRIMITIVES_SEAL_RESOURCES_RESOURCES_HPP
-#define CPP_FILECOIN_CORE_PRIMITIVES_SEAL_RESOURCES_RESOURCES_HPP
+#ifndef CPP_FILECOIN_CORE_PRIMITIVES_RESOURCES_RESOURCES_HPP
+#define CPP_FILECOIN_CORE_PRIMITIVES_RESOURCES_RESOURCES_HPP
 
 #include <boost/optional.hpp>
 #include <map>
@@ -374,4 +374,4 @@ namespace fc::primitives {
 
 }  // namespace fc::primitives
 
-#endif  // CPP_FILECOIN_CORE_PRIMITIVES_SEAL_RESOURCES_RESOURCES_HPP
+#endif  // CPP_FILECOIN_CORE_PRIMITIVES_RESOURCES_RESOURCES_HPP
