+++ core/sector_storage/CMakeLists.txt
@@ -24,4 +24,5 @@ add_library(scheduler
 
 target_link_libraries(scheduler
         outcome
+        resources
         )
